back/requirements.txt
flask
requests
python-dotenv
supabase
selenium
webdriver-manager